Python implementation of linked list
The linked list is also circular and doubly, which means
it can be navigated in both directions and will loop back
to the beginning when reached the end or to the end if reached the beginning.

You can add data to the beginning or to the end with append and prepend. The prepend takes
an additional argument called 'reset_current' and will reset the head to the prepended data
if true, otherwise it will do nothing and current will not be set to the first node in the linked list.
This could be useful if you want to add data to the linked list, but not change the current location in
the linked list.

You can get the data from the current node with the current method and you can change node with
next and previous methods.